Rename updateStoryNote util
