+++ packages/story-preview/utils/updateStoryNote.js
